+++ lib/src/types/types.dart
@@ -2,3 +2,4 @@ export 'year_month.dart';
 export 'gregorian_year.dart';
 export 'gregorian_month.dart';
 export 'gregorian_day.dart';
+export 'gregorian_month_day.dart';