+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 - Implement `xsd:nonPositiveInteger` data type codec
 - Implement `xsd:unsignedLong` data type codec
 - Implement `xsd:gDay` codec and data type
+- Implement `xsd:gMonthDay` codec and data type